add FindEventsByCalName
