+++ cmd/export-by-location/main.go
@@ -1,146 +1,34 @@
 package main
 
 import (
-	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"log"
-	"net/http"
-	"os"
 	"time"
 
-	"golang.org/x/oauth2"
-	"golang.org/x/oauth2/google"
-	"google.golang.org/api/calendar/v3"
-	"google.golang.org/api/option"
+	"google-calendar-utils/internal/agent"
+	"google-calendar-utils/pkg/calendar"
 )
-
-// Retrieve a token, saves the token, then returns the generated client.
-func getClient(config *oauth2.Config) *http.Client {
-	// The file token.json stores the user's access and refresh tokens, and is
-	// created automatically when the authorization flow completes for the first
-	// time.
-	tokFile := "token.json"
-	tok, err := tokenFromFile(tokFile)
-	if err != nil {
-		tok = getTokenFromWeb(config)
-		saveToken("secrets/"+tokFile, tok)
-	}
-	return config.Client(context.Background(), tok)
-}
-
-const (
-	localAdd           = ":8989"
-	redirectUriPattern = "/redirect"
-	queryParamName     = "code"
-)
-
-// Request a token from the web, then returns the retrieved token.
-func getTokenFromWeb(config *oauth2.Config) *oauth2.Token {
-	authURL := config.AuthCodeURL("state-token", oauth2.AccessTypeOffline)
-	fmt.Printf("Go to the following link in your browser then type the "+
-		"authorization code: \n%v\n", authURL)
-
-	authCode, err := getAuthCodeQueryParam(localAdd, redirectUriPattern, queryParamName)
-	if err != nil {
-		log.Fatalf("Unable to get auth code: %v", err)
-	}
-
-	tok, err := config.Exchange(context.TODO(), authCode)
-	if err != nil {
-		log.Fatalf("Unable to retrieve token from web: %v", err)
-	}
-	return tok
-}
-
-func getAuthCodeQueryParam(localAdd string, uriPattern string, paramName string) (string, error) {
-	var authCode string
-
-	codeChan := make(chan string)
-	errChan := make(chan error)
-	http.HandleFunc(uriPattern, func(w http.ResponseWriter, r *http.Request) {
-		queryVals := r.URL.Query()
-		codeChan <- queryVals.Get(paramName)
-	})
-	go func() {
-		err := http.ListenAndServe(localAdd, nil)
-		if err != nil {
-			errChan <- fmt.Errorf("Unable to launch local server: %w", err)
-		}
-	}()
-
-	select {
-	case authCode = <-codeChan:
-		if authCode == "" {
-			return "", errors.New("Unable to get auth code")
-		}
-		return authCode, nil
-	case err := <-errChan:
-		return "", err
-	}
-}
-
-// Retrieves a token from a local file.
-func tokenFromFile(file string) (*oauth2.Token, error) {
-	f, err := os.Open(file)
-	if err != nil {
-		return nil, err
-	}
-	defer f.Close()
-	tok := &oauth2.Token{}
-	err = json.NewDecoder(f).Decode(tok)
-	return tok, err
-}
-
-// Saves a token to a file path.
-func saveToken(path string, token *oauth2.Token) {
-	fmt.Printf("Saving credential file to: %s\n", path)
-	f, err := os.OpenFile(path, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0600)
-	if err != nil {
-		log.Fatalf("Unable to cache oauth token: %v", err)
-	}
-	defer f.Close()
-	json.NewEncoder(f).Encode(token)
-}
 
 func main() {
-	ctx := context.Background()
-	b, err := os.ReadFile("secrets/credentials.json")
-	if err != nil {
-		log.Fatalf("Unable to read client secret file: %v", err)
-	}
+	// TODO: コマンドライン引数を受け取る処理
+	calendarName := "KC111"
 
-	// If modifying these scopes, delete your previously saved token.json.
-	config, err := google.ConfigFromJSON(b, calendar.CalendarReadonlyScope)
-	if err != nil {
-		log.Fatalf("Unable to parse client secret file to config: %v", err)
-	}
-	client := getClient(config)
-
-	srv, err := calendar.NewService(ctx, option.WithHTTPClient(client))
+	srv, err := calendar.NewService("secrets/credentials.json")
 	if err != nil {
 		log.Fatalf("Unable to retrieve Calendar client: %v", err)
 	}
 
-	calendarList, err := srv.CalendarList.List().Do()
-	if err != nil {
-		log.Fatalf("Unable to retrieve next ten of the user's events: %v", err)
+	req := agent.FindRequest{
+		TimeMin: time.Date(2020, time.April, 1, 0, 0, 0, 0, time.FixedZone("JST", 0)),
+		TimeMax: time.Date(2023, time.April, 1, 0, 0, 0, 0, time.FixedZone("JST", 0)),
 	}
-	var targetId string
-	for _, cal := range calendarList.Items {
-		if cal.Summary == "KC111" {
-			targetId = cal.Id
-		}
-	}
-	fmt.Printf("KC111 calendarId: %s", targetId)
 
-	t := time.Date(2022, time.April, 1, 0, 0, 0, 0, time.FixedZone("JST", 0)).Format(time.RFC3339)
-	events, err := srv.Events.List(targetId).ShowDeleted(false).
-		SingleEvents(true).TimeMin(t).MaxResults(1000).OrderBy("startTime").Q("KC111").Do()
+	agent := agent.NewAgent(srv)
+	events, err := agent.FindEventsByCalName(calendarName, req)
 	if err != nil {
-		log.Fatalf("Unable to retrieve next ten of the user's events: %v", err)
+		log.Fatalf("Unable to find events by %s %v", calendarName, err)
 	}
+
 	fmt.Println("Upcoming events:")
 	if len(events.Items) == 0 {
 		fmt.Println("No upcoming events found.")
